Added python files, needs venv
Added all the files and output results from the function of the program. 

You will need to download your own virtual environment or libraries for the files to work. 

File was too big to bring in on one download had to break into each folder and the Venv(virtual environment) was too large (includes PIP and Python 3.12 interpreter)